requirements.txt
requests
openlineage-python
pytest

